SystemUI: Misc fixes for right (custom) lockscreen shortcuts
In the current implementation only the left target is being properly
updated via 'updateLeftAffordance'. Implement 'updateRightAffordance'
to handle the right target in a similar fashion.

This fixes (1) a rare scenario where the right target would be colored
black instead of gray and (2) the launch animation background for
custom targets not adopting the icon color of the selected app.

Change-Id: I556fc7316c3b646bf004b4f0a510fbc3736b2bd8